Move response and misfit views in separate files
